% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macrounchained.r
\name{macrounchainedFocusGW_ui}
\alias{macrounchainedFocusGW_ui}
\title{Text User Interface for macrounchainedFocusGW}
\usage{
macrounchainedFocusGW_ui()
}
\value{
See \code{\link[macrounchained:macrounchainedFocusGW]{macrounchainedFocusGW-methods}}.
}
\description{
Text User Interface for 
 \code{\link[macrounchained:macrounchainedFocusGW]{macrounchainedFocusGW-methods}}.
 The interface guide the user through a series of questions, 
 for fetching template Excel files for \code{macrounchainedFocusGW} 
 or picking an Excel file with parameter sets and starting 
 \code{macrounchainedFocusGW}.
}
