% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macrounchained.r
\docType{methods}
\name{macroutilsFocusGWConc_summary}
\alias{macroutilsFocusGWConc_summary}
\alias{macroutilsFocusGWConc_summary.list}
\title{Summarise one or several output results from macroutilsFocusGWConc (package macroutils2)}
\usage{
macroutilsFocusGWConc_summary(x, ...)

\method{macroutilsFocusGWConc_summary}{list}(x, info, ...)
}
\arguments{
\item{x}{A \code{\link{list}}, where each item is the output from 
\code{\link[macroutils2:macroutilsFocusGWConc-methods]{macroutilsFocusGWConc}}.}

\item{\dots}{Additional parameters passed to specific methods. 
Currently not used.}

\item{info}{A \code{\link[base]{data.frame}} with as many rows as 
items in \code{x}. Expected columns are \code{id} (the 
substance identifier), \code{run_id} (the simulation 
identifier) and optionally \code{name} (the name of the 
substance).}
}
\value{
A \code{\link[base]{data.frame}} with selected output 
 variables from \code{x} merged with \code{info}. Items 
 in \code{x} that are \code{NULL} are not retained.
}
\description{
Summarise one or several output results from 
 \code{\link[macroutils2:macroutilsFocusGWConc-methods]{macroutilsFocusGWConc}} 
 (package macroutils2). The functions is designed to be 
 passed to \code{\link[macrounchained:macrounchained-methods]{macrounchained}}, 
 via the argument \code{analyse_summary}.
}
