% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macrounchained.r
\docType{methods}
\name{macrounchained}
\alias{macrounchained}
\alias{macrounchained.data.frame}
\title{Batch run MACRO simulations with different substance properties and application patterns}
\usage{
macrounchained(s, ...)

\method{macrounchained}{data.frame}(s, parfile, verbose = 1L,
  indump = TRUE, run = TRUE, overwrite = FALSE, analyse = NULL,
  analyse_args = NULL, analyse_summary = NULL, dt50_depth_f = NULL,
  keep0conc = TRUE, anonymise = TRUE, archive = TRUE, ...)
}
\arguments{
\item{s}{A \code{\link[base]{data.frame}} containing different sets 
of substance properties and application patterns to be 
simulated.}

\item{\dots}{Additional parameters passed to specific methods. 
Currently not used.}

\item{parfile}{A \code{macroParFile}, as imported with 
\code{\link[rmacrolite]{rmacroliteImportParFile-methods}}}

\item{verbose}{Single integer value. If set to a value \code{< 1}, 
the program is silent (except on errors or warnings). If 
set to \code{1}, the program outputs messages. Values 
\code{> 1} may also activate messages from lower level 
functions (for debugging purpose).}

\item{indump}{Single logical value. If \code{TRUE} (the default), 
the so called \code{indump.tmp} parameter file is produced. 
Must be \code{TRUE} when \code{run} is \code{TRUE}.}

\item{run}{Single logical value. If \code{TRUE} (the default), 
the parametrised simulations are run.}

\item{overwrite}{Single logical value. If \code{FALSE} (the default), 
the function will check that the files to be created 
do not exist yet, and will \code{\link[base]{stop}} if 
some of the files already exist. Set to \code{TRUE} to 
silently overwrite existing files.}

\item{analyse}{Single R \code{\link[base]{function}}. Function to be 
used by \code{macrounchained} to analyse the results of 
MACRO simulations. An example of such function is 
\code{\link[macroutils2:macroutilsFocusGWConc-methods]{macroutilsFocusGWConc}}.
Notice that the appropriate function depends on what 
output needs to be analysed and what parameters are 
exported from MACRO, as defined in \code{parfile}, so 
there is no generic all purpose function to be used here.
When \code{analyse} is \code{NULL} (the default), MACRO 
output is not analysed.}

\item{analyse_args}{A \code{\link[base]{list}} containing named-items to 
be passed as arguments to \code{analyse}. An example 
of use is \code{analyse_args =} \code{list( "quiet" = TRUE )}.}

\item{analyse_summary}{Single R \code{\link[base]{function}}. Function to be 
used by \code{macrounchained} to summarise all the results 
of MACRO simulations, as output by \code{analyse}.
An example of such function is 
\code{\link[macrounchained:macroutilsFocusGWConc_summary-methods]{macroutilsFocusGWConc_summary}}.
Notice that the appropriate function depends on what 
output needs to be summarised and what output \code{analyse} 
is returning, so there is no generic all purpose function 
to be used here.}

\item{dt50_depth_f}{See \code{\link[rmacrolite:rmacroliteDegradation-methods]{rmacroliteDegradation}}.}

\item{keep0conc}{See \code{\link[rmacrolite:rmacroliteApplications-methods]{rmacroliteApplications}}.}

\item{anonymise}{Single boolean value. If \code{TRUE}, the function 
tries to replace USERNAME, HOMEPATH and USERPROFILE 
(i.e Windows environment variables) in paths displayed 
in messages by their environment variables, in order 
to preserve the user anonymity in the logs produced.}

\item{archive}{Single boolean value. If \code{TRUE}, all the files 
generated are archived in a \code{.tar.gz}-file when 
all other operations are finished. Default to 
\code{FALSE}.}
}
\value{
TO BE WRITTEN.
}
\description{
Batch run MACRO simulations with different substance 
 properties and application patterns, starting from a 
 template imported MACRO par-file, including metabolites.
}
